driver tweaks

+++ include/ros2_control_serial_diffdrive_plugin/serial_comms.h
@@ -20,7 +20,7 @@ public:
   void setup(const std::string &serial_device, int32_t baud_rate, int32_t timeout_ms);
   void activateMotors();
   void deactivateMotors();
-  void readEncoderValues(int &val_1, int &val_2);
+  void readEncoderValues(long int &val_1, long int &val_2);
   void setMotorValues(int val_1, int val_2);
   void setPidValues(float k_p, float k_d, float k_i, float k_o);
