+++ src/serial_comms.cpp
@@ -24,7 +24,7 @@ void SerialComms::deactivateMotors()
     std::string response = sendMsg("d\r", false);
 }
 
-void SerialComms::readEncoderValues(int &val_1, int &val_2)
+void SerialComms::readEncoderValues(long int &val_1, long int &val_2)
 {
     std::string response = sendMsg("e\r", true);
 
@@ -33,8 +33,8 @@ void SerialComms::readEncoderValues(int &val_1, int &val_2)
     std::string token_1 = response.substr(0, del_pos);
     std::string token_2 = response.substr(del_pos + delimiter.length());
 
-    val_1 = std::atoi(token_1.c_str());
-    val_2 = std::atoi(token_2.c_str());
+    val_1 = std::atol(token_1.c_str());
+    val_2 = std::atol(token_2.c_str());
 }
 
 void SerialComms::setMotorValues(int val_1, int val_2)
